feat(crud_page): split crud and crud view file and crud function
